Final Commit before Presentation
Using 3 separate socket transfers instead of string concat
Added global imgResult in handleButtonPush due to variable not defined error
Added .check_portion() method in drinksSDK to avoid pre-decrement drinks before they were actually used, which resulted in the program emptying the drinks in the database before running the program.
Changed the find_circle() function so that it returns a list if it finds a circle, and False if it does not. This allows the program to check for circles and not run unless there is a "glass" in the viewarea of the camera.